requirements.txt
requests>=2.28.0,<3.0.0
urllib3>=1.26.0,<3.0.0
pysnmp>=4.4.12,<6.0.0
PyYAML>=6.0,<7.0.0
ipaddress>=1.0.23;
python_version<"3.3"
